feat(pagination): improve usage

+++ client/src/main/java/dev/caceresenzo/gitbook/client/GitBookClient.java
@@ -69,7 +69,7 @@ public interface GitBookClient {
 	public static class Builder {
 
 		public static final String DEFAULT_API_URL = "https://api.gitbook.com";
-		public static final long DEFAULT_MAX_PAGE_SIZE = 1000;
+		public static final int DEFAULT_MAX_PAGE_SIZE = 1000;
 
 		/** The URL of the GitBook API. Defaults to `https://api.gitbook.com`. */
 		private String apiUrl = DEFAULT_API_URL;
@@ -78,7 +78,7 @@ public interface GitBookClient {
 		private String accessToken;
 
 		/** The page size used for pagination. */
-		private long maxPageSize = DEFAULT_MAX_PAGE_SIZE;
+		private int maxPageSize = DEFAULT_MAX_PAGE_SIZE;
 
 		public Builder unauthenticated() {
 			return accessToken(null);
